NOT WORKING
- Removed an HGE macro that was causing VAX to have kittens.
- Updated RakNet to 3.25.
- Matchmaking now uses a proper POST stream not an octet-stream.
- Enabled latency simulation for debug builds (80-120).
- Started adding full Send()/Broadcast() capabilities to both clients and host.

git-svn-id: svn+ssh://supergirl.dreamhost.com/home/sulfate/svn/pikpik@106 d6e08b45-3f4a-0410-be67-99b8f2306d8d

+++ RakNet/PacketFileLogger.cpp
@@ -18,9 +18,9 @@ void PacketFileLogger::StartLog(const char *filenamePrefix)
 	// Open file for writing
 	char filename[256];
 	if (filenamePrefix)
-		sprintf(filename, "%s_%i.csv", filenamePrefix, RakNet::GetTime());
+		sprintf(filename, "%s_%i.csv", filenamePrefix, (int) RakNet::GetTime());
 	else
-		sprintf(filename, "PacketLog_%i.csv", RakNet::GetTime());
+		sprintf(filename, "PacketLog_%i.csv", (int) RakNet::GetTime());
 	packetLogFile = fopen(filename, "wt");
 	LogHeader();
 	if (packetLogFile)
